requirements.txt
africastalking==1.2.9
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.5.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.3
django-appconf==1.1.0
django-compressor==4.5.1
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-environ==0.11.2
django-imagekit==5.0.0
django-ipware==7.0.1
django-js-asset==3.0.1
django-mptt==0.16.0
django-redis==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.4
packaging==24.2
pilkit==3.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-ipware==3.0.0
pytz==2024.2
PyYAML==6.0.2
rcssmin==1.1.2
redis==6.4.0
referencing==0.35.1
requests==2.32.3
resend==2.4.0
rjsmin==1.2.2
rpds-py==0.22.3
schema==0.7.7
secure-smtplib==0.1.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
